Chaning the colors parsing pointers from float to double to fix the deprecation warnings for iOS 8

+++ gems/under-os-core/lib/under_os/color.rb
@@ -29,10 +29,10 @@ class UnderOs::Color
   end
 
   def to_rgba
-    r = Pointer.new(:float)
-    g = Pointer.new(:float)
-    b = Pointer.new(:float)
-    a = Pointer.new(:float)
+    r = Pointer.new(:double)
+    g = Pointer.new(:double)
+    b = Pointer.new(:double)
+    a = Pointer.new(:double)
 
     @ui.getRed(r, green:g, blue:b, alpha:a)
